requirements-flask3.txt
Flask==3.0.2
DateTime==5.3
email-validator==2.1.0.post1
exceptiongroup==1.2.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-sqlalchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
Jinja2==3.1.3
lxml==4.9.4
pysftp==0.2.9
paramiko==3.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.25
Werkzeug==3.0.1
WTForms==3.1.2
xmltodict==0.13.0